requirements.txt
pip~=23.2.1
wheel~=0.41.2
PyYAML~=6.0.1
setuptools~=68.2.0
pandas~=2.1.3
matplotlib~=3.8.2
scikit-learn~=1.3.2
exceptiongroup~=1.2.0
trio==0.22.0
anyio==3.6.2
asks>=3.0.0
requests>=2.28.2
python-qpid-proton
python-dotenv